Add animation to entity deletion
